pyproject.toml
[project]
name = "discord-acp-bridge"
version = "0.1.0"
description = "Discord Bridge for ACP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.0.0",
    "agent-client-protocol>=0.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.0",
    "mypy>=1.19.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/discord_acp_bridge"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "discord.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "acp.*"
ignore_missing_imports = true
